melhor detalhe para o boiler

+++ firmware/common/hal.h
@@ -19,7 +19,7 @@ uint8_t p1(uint8_t state);
 // tanque 3
 uint8_t s31();
 uint8_t s32();
-uint8_t bs1();
+uint32_t bs1();
 uint8_t v2(uint8_t state);
 uint8_t b1(int8_t state);
